// appflor-fe/src/components/Arreglos/Types/ArregloTypes.ts
export type Arreglo = {
  id_arreglo?: number;
  nombre: string;
  fecha?: string;
  descripcion: string;
  precio: number;
  categoria: string;
  imagen?: string;
  usuarioCreacion?: string;
  fechaCreacion?: string;
  usuarioModificacion?: string;
  fechaModificacion?: string;
  estado?: string;
};